perf: constrain main bits in equality diff to be only zeros
When splitting the diff into main and carry part, we checked that the main part
is zero and carried the carry part separately to next iteration. Because we
have a special case (main part == 0 iff all bits are zeros), then it is more
beneficial to just constrain every bit to be zero.